requirements.txt
asgiref==3.2.10
attrs==20.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-easy-pdf==0.1.1
django-extra-views==0.13.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
freezegun==1.0.0
gunicorn==20.0.4
html5lib==1.1
importlib-metadata==2.0.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.9.5
pytest==6.1.1
pytest-django==3.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
reportlab==3.5.53
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
zipp==3.3.0
